feat: frontend draws between players, fixed infinite rerendering on the game form caused by always changing table data in the sortable context
