+++ magicstats.client/src/Games/GameForm/DragAndDropParticipantsList.css
@@ -5,3 +5,6 @@
     cursor: pointer;
     opacity: 80%;
 }
+.dnd-participants-list {
+    width: 100%;
+}
